+++ pysat/instruments/templates/netcdf_pandas.py
@@ -22,11 +22,10 @@ easy.
 
 """
 
-import pandas as pds
-import numpy as np
+import datetime as dt
+import logging
 import pysat
 
-import logging
 logger = logging.getLogger(__name__)
 
 # pysat required parameters
@@ -36,7 +35,7 @@ name = 'pandas'
 tags = {'': 'netCDF4'}
 # dictionary of satellite IDs, list of corresponding tags
 sat_ids = {'': ['']}
-_test_dates = {'': {'': pysat.datetime(2009, 1, 1)}}
+_test_dates = {'': {'': dt.datetime(2009, 1, 1)}}
 
 
 def init(self):
@@ -210,6 +209,6 @@ def download(date_array, tag, sat_id, data_path=None, user=None,
 
     """
 
-    logger.info('This is a generic Instrument routine and does not support ' +
-          'downloading data.')
+    logger.info(' '.join(('This is a generic Instrument routine and does not',
+                          'support downloading data.')))
     pass