+++ CHANGELOG.md
@@ -2,23 +2,34 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [3.0.0] - 2020-03-07
+## [3.0.0] - 2020-04-24
 - New Features
   - Added registry module for registering custom external instruments
   - Added Meta.mutable flag to control attribute mutability
   - custom.attach replaces custom.add
   - Unit tests are now pytest compatible
+  - Added altitudes to test instruments
+  - New flags added to instruments to streamline unit testing: `_test_download`, `_test_download_travis`, `_password_req`
+  - Madrigal instruments migrated to pysatMadrigal
+  - methods.nasa_cdaweb.list_files moved to methods.general
 - Deprecations
   - Removed ssnl
   - Removed utils.stats
   - Removed model_utils
+  - Removed deprecated pandas.Panel
+  - imports datetime from datetime, not pandas (deprecated)
+  - import DataFrame and Series directly from pandas, not pysat
   - Removed coords.scale_units
   - Removed time.season_date_range
+  - DeprecationWarning for strict_time_flag only triggered if sloppy data is found
 - Documentation
   - Added info on how to register new instruments
   - Fixed description of tag and sat_id behaviour in testing instruments
 - Bug Fix
   - Fixed custom instrument attribute persistence upon load
+- Maintenance
+  - nose dependency removed from unit tests
+  - Specify dtype for empty pandas.Series for forward compatibility
 
 ## [2.2.0] - 2020-2-29
 - New Features
@@ -49,6 +60,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Added small time offsets (< 1s) to ensure COSMIC files and data have unique times
   - Updates to Travis CI environment
   - Removed `inplace` use in xarray `assign` function, which is no longer allowed
+  - Fixed output of orbit_info during print(inst)
 
 
 ## [2.1.0] - 2019-11-18